Fix issues from review

+++ core/server/src/prover_server/scaler.rs
@@ -17,17 +17,16 @@ impl ScalerOracle {
         Self { db }
     }
 
-    /// Decides how many prover entities should be created depending on the amount of existing
-    /// entities and the amount of pending blocks.
+    /// Decides how many prover entities should be created depending on the amount of pending blocks.
     pub fn provers_required(&mut self) -> Result<u32, failure::Error> {
         // Currently the logic of this method is very simple:
         // We require a prover for each pending block plus IDLE_RROVERS for faster upscaling.
 
         let storage = self.db.access_storage()?;
         let pending_jobs = storage.prover_schema().pending_jobs_count()?;
-        let idle_provers: u64 = parse_env("IDLE_PROVERS");
+        let idle_provers: u32 = parse_env("IDLE_PROVERS");
         let provers_required = pending_jobs + idle_provers;
 
-        Ok(provers_required as u32)
+        Ok(provers_required)
     }
 }
